requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2 